remove static field
